refactor: warnings in PplService

+++ RedEdrShared/logging.cpp
@@ -126,7 +126,7 @@ void LOG_W(int verbosity, const wchar_t* format, ...)
 void LOG_A(int verbosity, const char* format, ...)
 {
     char message[DATA_BUFFER_SIZE] = "[RedEdr PPL] ";
-    DWORD offset = strlen(message);
+    size_t offset = strlen(message);
 
     va_list arg_ptr;
     va_start(arg_ptr, format);
@@ -140,7 +140,7 @@ void LOG_A(int verbosity, const char* format, ...)
 void LOG_W(int verbosity, const wchar_t* format, ...)
 {
     WCHAR message[WCHAR_BUFFER_SIZE] = L"[RedEdr PPL] ";
-    DWORD offset = wcslen(message);
+    size_t offset = wcslen(message);
 
     va_list arg_ptr;
     va_start(arg_ptr, format);
@@ -170,7 +170,7 @@ void LOG_W(int verbosity, const wchar_t* format, ...)
 void LOG_A(int verbosity, const char* format, ...)
 {
     char message[DATA_BUFFER_SIZE] = "[RedEdr PPL] ";
-    DWORD offset = strlen(message);
+    size_t offset = strlen(message);
 
     va_list arg_ptr;
     va_start(arg_ptr, format);
@@ -184,7 +184,7 @@ void LOG_A(int verbosity, const char* format, ...)
 void LOG_W(int verbosity, const wchar_t* format, ...)
 {
     WCHAR message[WCHAR_BUFFER_SIZE] = L"[RedEdr PPL] ";
-    DWORD offset = wcslen(message);
+    size_t offset = wcslen(message);
 
     va_list arg_ptr;
     va_start(arg_ptr, format);
